Rework the Provider and Parser to allow any value to be used
The main drawback of the initial Provider/Parser design was that the
raw values exchanged between the interfaces were limited to strings,
thanks to the original design for CLI arguments and environment
variables. This obviously is a problem for more advanced providers that
could be added later, like JSON, YAML.

This change the type of the exchange variables to interface and rewrite
the standard parser to parse from string to the types understood by the
old parser, which will allow extension in the future.