Fix anonymous field marking
The anonymous status was previously detected using a function that has
been removed in favor of an attribute on the `Field` struct. When doing
that we forgot to account for the root field, which in turn prevented
anything else from being configured.

This fix the attribute setting and add a relevant test to detect this.