Add CustomParser interface to allow custom type to be use in ParseString
Adding a CustomParser allow a client to implement the interface to use custom type
for example a typed string.

ParseString is able to check if the res variable implement the CustomParser interface then call it.
It returns an error if the method Parse is failing.
If the res variable do not implement the CustomParser interface the normal behavior will returns an error.